# Conflicts:
#	projects/patchouli_datagen
#	projects/simplenetlib
#	src/main/generated/.cache/cache
#	src/main/java/com/github/minecraftschurlimods/arsmagicalegacy/common/util/AMUtil.java

+++ src/main/java/com/github/minecraftschurlimods/arsmagicalegacy/compat/patchouli/MainBlockStateMatcher.java
@@ -17,8 +17,8 @@ class MainBlockStateMatcher implements IStateMatcher {
     }
 
     @Override
-    public BlockState getDisplayedState(int ticks) {
-        AltarStructureMaterial mat = AltarMaterialManager.instance().getRandomStructureMaterial(ticks / 20);
+    public BlockState getDisplayedState(long ticks) {
+        AltarStructureMaterial mat = AltarMaterialManager.instance().getRandomStructureMaterial(Math.toIntExact(ticks / 20));
         if (mat == null) return Blocks.AIR.defaultBlockState();
         return mat.block().defaultBlockState();
     }
